+++ app/Main.hs
@@ -2,6 +2,7 @@ module Main (main) where
 
 import           ClassyPrelude
 
+import           Control.Monad.ST      (runST)
 import qualified Data.ByteString.Char8 as BC
 import           Options.Applicative   (auto, execParser, flag', fullDesc, header, help,
                                         helper, info, long, option, optional, progDesc,
@@ -10,6 +11,7 @@ import           System.Directory      (doesFileExist, getFileSize)
 
 
 import           Template              (spiralWorldTemplate, zigZagWorldTemplate)
+import qualified Vector                as VecWorld
 import           World                 (ProcessingKind (..), addCoords, findLargestIsland,
                                         mkWorld, mkWorldByTemplate, parseWorld)
 
@@ -17,7 +19,8 @@ import           World                 (ProcessingKind (..), addCoords, findLarg
 data ProcessType = Spiral        -- ^ use SpiralWorld template
                  | ZigZag        -- ^ use ZigZagWorld template
                  | NoTemplate    -- ^ find coordinates in SpiralWorld by algorithm
-  deriving (Show)
+                 | Vector        -- ^ find coordinates in SpiralWorld by algorithm
+  deriving (Show, Eq)
 
 
 -- | Any cmd line arguments overrides config
@@ -40,18 +43,25 @@ main = do
   doesFileExist argScrollPath >>= \case
     False -> putStrLn noFileError
     True  -> do
-      size <- fromIntegral <$> getFileSize argScrollPath
-      pw   <- parseWorld <$> BC.readFile argScrollPath
+      size   <- fromIntegral <$> getFileSize argScrollPath
+      scroll <- BC.readFile argScrollPath
 
-      let procKind = case argChunkSize of
+      let pw       = parseWorld scroll
+          procKind = case argChunkSize of
                        Nothing    -> Sequential
                        Just chunk -> Parallel chunk
-          vlgs = case argProcessType of
-                   Spiral     -> addCoords $ mkWorldByTemplate size spiralWorldTemplate pw
-                   ZigZag     -> addCoords $ mkWorldByTemplate size zigZagWorldTemplate pw
-                   NoTemplate -> mkWorld size procKind pw
+          maxPop = case argProcessType of
+                     Spiral     -> findLargestIsland
+                                 . addCoords
+                                 $ mkWorldByTemplate size spiralWorldTemplate pw
+                     ZigZag     -> findLargestIsland
+                                 . addCoords
+                                 $ mkWorldByTemplate size zigZagWorldTemplate pw
+                     NoTemplate -> findLargestIsland
+                                 $ mkWorld size procKind pw
+                     Vector     -> runST (VecWorld.parseAndProcess size scroll)
 
-      putStrLn $ "Largest population on an island is: " <> tshow (findLargestIsland vlgs)
+      putStrLn $ "Largest population on an island is: " <> tshow maxPop
 
   where
     pInfo = info (args <**> helper)
@@ -60,7 +70,7 @@ main = do
     args = Args <$> strOption (long "scroll"  <> short 's'
                             <> value "data/the.scroll" <> showDefault
                             <> help "Path of the scroll file")
-                <*> (spiralP <|> zigZagP <|> noTemplateP)
+                <*> (spiralP <|> zigZagP <|> noTemplateP <|> vectorP)
                 <*> optional (option auto (long "chunk_size" <> short 'z'
                                         <> help "Size of chunk in bytes for concurrent parse")
                              )
@@ -68,6 +78,7 @@ main = do
     spiralP     = flag' Spiral     (long "spiral"      <> help "Spiral template")
     zigZagP     = flag' ZigZag     (long "zig-zag"     <> help "Zig-zag template")
     noTemplateP = flag' NoTemplate (long "no-template" <> help "No template, Spiral")
+    vectorP     = flag' Vector     (long "by-vector"   <> help "Using vector, no coordinates, Spiral")
 
 
 noFileError :: Text